+++ src/globcor.h
@@ -9,14 +9,14 @@
 #define MELT_MOD_GLOBCOR_H_
 
 void jd2monat();
-double deklination(long tag, long monat);
-double gettimeequation(long tag,long monat);
-double sonnenhoehe(long tag,long monat,double breite,double laenge,double stunde);
-void getsuncoordinates(long tag,long monat,double breite,double laenge,double stunde,
+double deklination(int tag, int monat);
+double gettimeequation(int tag,int monat);
+double sonnenhoehe(int tag,int monat,double breite,double laenge,double stunde);
+void getsuncoordinates(int tag,int monat,double breite,double laenge,double stunde,
                        double *sunazimuth,double *sunaltitude);
 
 double getcosincidence(double aspect,double slope,double azimuth,double zenith);
-void sunrise(long tag,long monat,double breite,double laenge,double *sr,double *ss);
+void sunrise(int tag,int monat,double breite,double laenge,double *sr,double *ss);
 void radiussunearth();
 void schatten();
 void getzenith();
